+++ playwright-e2e/fixtures/create-fixture.ts
@@ -56,6 +56,8 @@ import { ManageCaseDocumentsPage } from '../pages/events/manage-case-documents/M
 import { CheckYourAnswersPage } from '../pages/CheckYourAnswersPage';
 import {AddNotePage} from '../pages/events/add-note/AddNotePage.ts';
 import { AllocateToJudgePage } from '../pages/events/allocate-to-judge/AllocateToJudgePage';
+import {EventSummaryPage} from "../pages/events/EventSummaryPage.ts";
+import {RefundPage} from "../pages/events/refund/RefundPage.ts";
 
 const commonActionsHelper = new CommonActionsHelper();
 const solicitorDetailsHelper = new SolicitorDetailsHelper();
@@ -115,6 +117,8 @@ type CreateFixtures = {
   checkYourAnswersPage: CheckYourAnswersPage;
   addNotePage: AddNotePage;
   allocateToJudgePage: AllocateToJudgePage;
+  eventSummaryPage: EventSummaryPage;
+  refundPage: RefundPage;
 };
 
 export const test = base.extend<CreateFixtures>({
@@ -279,5 +283,11 @@ export const test = base.extend<CreateFixtures>({
   },
   allocateToJudgePage: async ({ page }, use) => {
     await use(new AllocateToJudgePage(page));
-  }
+  },
+  eventSummaryPage: async ({ page }, use) => {
+    await use(new EventSummaryPage(page));
+  },
+    refundPage: async ({ page }, use) => {
+        await use(new RefundPage(page));
+    }
 });
